yolo_web/requirements.txt
flask
torch
opencv-python
ultralytics
pillow